requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.5
et-xmlfile==1.1.0
future==0.18.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.2
jaraco.context==4.0.0
jsonschema==3.2.0
Logbook==1.5.3
matrix-nio==0.18.6
more-itertools==8.9.0
multidict==5.1.0
numpy==1.21.2
openai==0.10.4
openpyxl==3.0.8
pandas==1.3.3
pandas-stubs==1.2.0.16
peewee==3.14.4
pycountry==20.7.3
pycparser==2.20
pycryptodome==3.10.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-olm==3.1.3
python-socks==1.2.4
pytz==2021.1
requests==2.26.0
SimpleParse==2.2.3
six==1.16.0
tqdm==4.62.2
typing-extensions==3.10.0.2
unpaddedbase64==2.1.0
urllib3==1.26.6
wolframalpha==5.0.0
xmltodict==0.12.0
yarl==1.6.3
